old-tests/submission/Microsoft/video/video_007.htm

<!DOCTYPE HTML>
<html>
<head>
  <title>HTML5 Media Elements: Seeking to a value less than 'startTime' seeks to 'startTime'.</title>
  <meta content="text/html; charset=UTF-8" http-equiv="Content-Type">
  <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
  <link rel="help" href="http://www.w3.org/TR/html5/video.html#dom-media-seeking" />
  <meta name="assert" content="Seeking to a value less than 'initialTime' seeks to 'initialTime'." />
  <meta name=timeout content=long>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
  <script type="text/javascript" src="../common/media.js"></script>
  <script type="text/javascript">
	    var videotest = async_test("Seeking to a value less than 'initialTime' seeks to 'initialTime'.");

function getTestVideoElement()
{
    return document.getElementById("video0");
}

function do_play(event)
{
    var vid = getTestVideoElement();
    vid.pause();
    vid.currentTime = -1;
}

function do_seeked(event)
{
    videotest.step(function() {
      var vid = getTestVideoElement();
    
      assert_equals(vid.currentTime, vid.initialTime);
    });
    videotest.done();
}

</script>
</head>
<body>
<pre>
Seeking to a value less than 'initialTime' seeks to 'initialTime'.
</pre>
<pre>

http://www.w3.org/TR/html5/video.html#dom-media-seeking

When the user agent is required to seek to a particular new playback 
position in the media resource, it means that the user agent must run 
the following steps: 

[...]

3. If the new playback position is less than the earliest possible 
position, let it be that position instead. 

</pre>
<div id='log'></div>

<script type="text/javascript">

document.write(
    "<video id='video0' src='" + getVideoURI("/media/movie_5") + "'" +
    " onseeked='do_seeked(event);'" +
    " onplay='do_play(event);'" +
    " autoplay>"
    );

</script>
Your browser does not support media elements.
</video>

</body>
</html>
